data: Track: chop withMainRelease
